requirements.txt
streamlit==1.26.0
pillow==9.4.0
numpy==1.23.5
opencv-python==4.8.0.74
opencv-contrib-python==4.8.0.74
